refactor: relocate SourceCard to artist profile components
