+++ src/app/artist/[id]/_components/SourceCard.tsx
